refactor(tests): update index and helloTessera tests
